1. Fix bug for data-update after adding new data-node ※ cause : new-node does not have a collection's info
2. Fix bug for searching in new collection what is made by
collection-wizard
※ cause : loadbalancer does not contains new-collection's id

+++ server/src/main/java/org/fastcatsearch/ir/IRService.java
@@ -360,7 +360,9 @@ public class IRService extends AbstractService {
 				
 				FilePaths collectionFilePaths = environment.filePaths().collectionFilePaths(collectionId);
 				//FileUtils.deleteDirectory(collectionFilePaths.file());
-				FileUtils.forceDelete(collectionFilePaths.file());
+				if(collectionFilePaths.file().exists()) {
+					FileUtils.forceDelete(collectionFilePaths.file());
+				}
 				return true;
 			} catch (Exception e) {
 				logger.error("Error while remove collection", e);
